Fixes issue 47 (https://github.com/rainlab/blog-plugin/issues/47)
Intead of letting Parsedown interpret the shortcuts for inserting
images via Dropzone (e.g. ![1](image)), we parse them ourselves
before passing the input over to Parsedown.

Note: Classes and id:s can't be added with the MarkdownExtra
syntax as in the documentation. These have to be added *after*
the acutal image upload has been made. This is actually also
the case with the existing code, and should either be fixed or
mentioned in the documentation.