Improve support for Build 420
As of Laravel 5.2 the `nullable` validation rule is required to be added to properties that are allowed to be NULL. Side note: The `code` attribute on the Category model does not appear to be used anywhere or by anything. I suggest that it should be removed from the model and the DB schema. @daftspunk thoughts?