Remove check for category relation
The categories relation should always exist, as it is defined in the same model. However, the ``getRelations()`` method appears to be buggy, as it does not return all defined relations. For example when I install the ``Bedard.BlogTags`` plugin, this will cover all other relations and only the newly added ``tags`` relation is beein returned.