Rollback custom slugs
This is an aspect long considered during the design phase of the plugin and October in general, it made sense to keep some properties as constants for the sake of simplicity and execution; a careful decision weighed against the ability to comprehend the code and execution life-cycle that comes at the sacrifice of a small amount of flexibility. The gymnastics and technical debt that this feature adds is evident in the code that was rolled back. It would be worth exploring alternative options on how this can be solved without introducing this type of thing to the codebase

This appears to fix several third party plugins that depend on RainLab.Blog which is a key component

Refs #392
Refs #394
Rollback 950fbd7b1e6a0ad8d1c0277027fe3ed7c8e12091