Update Category slug in URL on language switch to match translation
When using Rainlab.Blog and Rainlab.Translate together the category slug part of the address bar URL is not updated when switching languages (using LangPicker component). Neither is a correct URL used in alternateHrefLangElements component.

This change allows for the category slug part of the URL (in blogPost URL as well as in blogCategory URL) to be updated/changed into its correct localised/translated version.